+++ Transmittal.Desktop/Helpers/FilenameParser.cs
@@ -1,53 +1,12 @@
 ﻿using Humanizer;
-using System.Diagnostics;
 using System.IO;
 using System.Text.RegularExpressions;
 using Transmittal.Library.Models;
 using Transmittal.Library.Extensions;
 
-namespace Transmittal.Desktop.Util;
-internal static class ISO19650Parser
+namespace Transmittal.Desktop.Helpers;
+internal static class FilenameParser
 {
-    internal static List<DocumentTypeModel> GetDocumentTypes()
-    {
-        List<DocumentTypeModel> documentTypes = new List<DocumentTypeModel>();
-
-        documentTypes.Clear();
-        documentTypes.Add(new DocumentTypeModel() { Code = "AF", Description = "Animation file (of a model)" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "CM", Description = "Combined model (combined multidiscipline model)" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "CR", Description = "Specific for the clash process" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "DR", Description = "2D drawing" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "M2", Description = "2D model file" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "M3", Description = "3D model file" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "MR", Description = "Model rendition file For other renditions, e.g thermal analysis etc." });
-        documentTypes.Add(new DocumentTypeModel() { Code = "VS", Description = "Visualization file (of a model)" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "BQ", Description = "Bill of quantities" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "CA", Description = "Calculations" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "CO", Description = "Correspondence" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "CP", Description = "Cost plan" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "DB", Description = "Database" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "FN", Description = "File note" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "HS", Description = "Health And safety" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "IE", Description = "Information exchange file" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "MI", Description = "Minutes / Action notes" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "MS", Description = "Method statement" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "PP", Description = "Presentation" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "PR", Description = "Programme" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "RD", Description = "Room data sheet" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "RI", Description = "Request for information" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "RP", Description = "Report" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "SA", Description = "Schedule of accommodation" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "SH", Description = "Schedule" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "SN", Description = "Snagging list" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "SP", Description = "Specification" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "SU", Description = "Survey" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "GA", Description = "General arrangement drawing" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "DT", Description = "Detail/assembly drawing" });
-        documentTypes.Add(new DocumentTypeModel() { Code = "SK", Description = "Sketch drawings" });
-
-        return documentTypes;
-    }
-
     internal static DocumentModel DocumentModel(string filePath, string projectIdentifier, string originator, string role)
     {
         string DocNo = "0000";
@@ -149,7 +108,7 @@ internal static class ISO19650Parser
         return document;
     }
 
-    internal static DocumentModel DocumentModel(string filePath, string projectIdentifier, string originator, string role, string exportRule)
+    public static DocumentModel DocumentModel(string filePath, string projectIdentifier, string originator, string role, string exportRule)
     {
         // Generate the regular expression pattern from the export rule
         var pattern = GetPatternFromExportRule(exportRule);
@@ -165,7 +124,7 @@ internal static class ISO19650Parser
         var documentProperties = new Dictionary<string, string>();
         foreach (Group group in match.Groups)
         {
-            documentProperties[group.Name] = group.Value;          
+            documentProperties[group.Name] = group.Value;
         }
 
         DocumentModel document = new DocumentModel
@@ -185,7 +144,6 @@ internal static class ISO19650Parser
 
         return document;
     }
-
 
     private static string GetPatternFromExportRule(string exportRule)
     {
