+++ Transmittal.Library/Transmittal.Library.csproj
@@ -21,7 +21,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="CommunityToolkit.Mvvm" Version="8.1.0-preview3" />
+    <PackageReference Include="CommunityToolkit.Mvvm" Version="8.1.0" />
     <PackageReference Include="Dapper" Version="2.0.123" />
     <PackageReference Include="Humanizer.Core" Version="2.14.1" />
 	<PackageReference Include="Microsoft.Data.Sqlite" Version="7.0.1" />
